chore: add test codes and fix ConfigCount
