--- Tests/ListKitTests/ListKitTests.swift
@@ -1,15 +1,0 @@
-@testable import ListKit
-import XCTest
-
-final class ListKitTests: XCTestCase {
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct
-        // results.
-        XCTAssertEqual("Hello, World!", "Hello, World!")
-    }
-
-    static var allTests = [
-        ("testExample", testExample),
-    ]
-}
